:recycle: Refactor shared utilities to real shared general purpose location
They're not limited to usage in registration backends, but
also useful for the payment plugin configuration, and, at
some point, probably also for prefill.
