+++ src/openforms/js/components/admin/form_design/payments/ogone_legacy/OgoneLegacyOptionsForm.js
@@ -2,16 +2,13 @@ import PropTypes from 'prop-types';
 import {useContext} from 'react';
 import {FormattedMessage} from 'react-intl';
 
-// TODO: move into more generic location
-import {
-  filterErrors,
-  getChoicesFromSchema,
-} from 'components/admin/form_design/registrations/shared/utils';
 import Fieldset from 'components/admin/forms/Fieldset';
 import {
   ValidationErrorContext,
   ValidationErrorsProvider,
+  filterErrors,
 } from 'components/admin/forms/ValidationErrors';
+import {getChoicesFromSchema} from 'utils/json-schema';
 
 import OptionsConfiguration from '../OptionsConfiguration';
 import {MerchantID} from './fields';
