+++ src/openforms/js/utils/json-schema.js
@@ -1,0 +1,8 @@
+export const getChoicesFromSchema = (enums, enumNames) => {
+  const finalChoices = [];
+  Object.keys(enums).forEach(key => {
+    finalChoices.push([enums[key], enumNames[key]]);
+  });
+
+  return finalChoices;
+};
